[nextest-runner] move PathMapper to its own module
Going to introduce more data structures to this module.
